+++ bidon/src/main/java/org/bidon/sdk/ads/Ad.kt
@@ -10,7 +10,6 @@ import java.util.*
  */
 class Ad(
     val demandAd: DemandAd,
-    val bidType: BidType,
     val ecpm: Double,
     val roundId: String,
     val auctionId: String,
@@ -22,7 +21,10 @@ class Ad(
     val networkName: String
         get() = adUnit.demandId
 
+    val bidType: BidType
+        get() = adUnit.bidType
+
     override fun toString(): String {
-        return "Ad(${demandAd.adType} $networkName/$bidType $ecpm $currencyCode, auctionId=$auctionId, roundId=$roundId, dsp=$dsp, extras=${demandAd.getExtras()})"
+        return "Ad(${demandAd.adType} $networkName/$bidType $ecpm $currencyCode, auctionId=$auctionId, roundId=$roundId, dsp=$dsp, extras=${demandAd.getExtras()}, $adUnit)"
     }
 }
