+++ bidon/src/main/java/org/bidon/sdk/stats/impl/StatisticsCollectorImpl.kt
@@ -94,7 +94,6 @@ class StatisticsCollectorImpl : StatisticsCollector {
             roundId = roundId,
             auctionId = auctionId,
             dsp = stat.dspSource,
-            bidType = bidType,
             adUnit = adUnit
         )
     }
